check notes
I think the issue in the code now is that showRejection, showAchievement, and showRandomEvent are all slightly different < this is a mistake.

I need to write one code that applies to all and null certain elements so it's easier to add on.

Right now, the totalRejection is based on a separate counter the jobApplication counter. Make it so that the totalRejection counter is based on the number of times showRejection triggers